Added the use of trajectories with stride.
I have added the use of trajectories for both
reference and deformed, and it will loop over
the deformed trajectory, with the reason being
that it would be more common to only have a
reference pdb and deformed trajectory than the
other way around. Ofc, it also works with just
pdb files. The one caveat is that I don't think
it works if you supply ref trajectory and NOT
a deformed trajectory. I will fix this in the
future, if needed.